Discourse: Refactor like weight logic
In preparation for adding more weight calculations for likes, weight
calculations are moved to their own file. This builds towards the
completion of #2455

n.b. This is similar to how the experimental discord reaction weight
calculations are organized.

test plan: yarn unit
